Extract IngestTaskRunner from IngestJobQueueConsumerRunner
