+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/job/IngestJobQueueConsumerRunner.java
@@ -30,10 +30,8 @@ import sleeper.configuration.jars.ObjectFactoryException;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.core.iterator.IteratorException;
-import sleeper.ingest.IngestResult;
 import sleeper.ingest.status.store.task.DynamoDBIngestTaskStatusStore;
-import sleeper.ingest.task.IngestTaskFinishedStatus;
-import sleeper.ingest.task.IngestTaskStatus;
+import sleeper.ingest.task.IngestTaskRunner;
 import sleeper.ingest.task.IngestTaskStatusStore;
 import sleeper.statestore.StateStoreException;
 import sleeper.statestore.StateStoreProvider;
@@ -41,55 +39,13 @@ import sleeper.utils.HadoopConfigurationProvider;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 
 import java.io.IOException;
-import java.time.Instant;
 import java.util.UUID;
-import java.util.function.Supplier;
 
 public class IngestJobQueueConsumerRunner {
+    private IngestJobQueueConsumerRunner() {
+    }
+
     private static final Logger LOGGER = LoggerFactory.getLogger(IngestJobQueueConsumerRunner.class);
-
-    private final IngestJobSource jobSource;
-    private final String taskId;
-    private final IngestTaskStatusStore statusStore;
-    private final Supplier<Instant> getTimeNow;
-    private final IngestJobHandler runJobCallback;
-
-    public IngestJobQueueConsumerRunner(
-            IngestJobSource jobSource, String taskId, IngestTaskStatusStore statusStore,
-            IngestJobHandler runJobCallback, Supplier<Instant> getTimeNow) {
-        this.getTimeNow = getTimeNow;
-        this.jobSource = jobSource;
-        this.taskId = taskId;
-        this.statusStore = statusStore;
-        this.runJobCallback = runJobCallback;
-    }
-
-    public IngestJobQueueConsumerRunner(
-            IngestJobQueueConsumer queueConsumer, String taskId, IngestTaskStatusStore statusStore, IngestJobHandler runJobCallback) {
-        this(queueConsumer, taskId, statusStore, runJobCallback, Instant::now);
-    }
-
-    public void run() throws IOException, IteratorException, StateStoreException {
-        Instant startTaskTime = getTimeNow.get();
-        IngestTaskStatus.Builder taskStatusBuilder = IngestTaskStatus.builder().taskId(taskId).startTime(startTaskTime);
-        statusStore.taskStarted(taskStatusBuilder.build());
-        LOGGER.info("IngestTask started at = {}", startTaskTime);
-
-        IngestTaskFinishedStatus.Builder taskFinishedStatusBuilder = IngestTaskFinishedStatus.builder();
-        jobSource.consumeJobs(job -> {
-            Instant startTime = getTimeNow.get();
-            IngestResult result = runJobCallback.ingest(job);
-            Instant finishTime = getTimeNow.get();
-            taskFinishedStatusBuilder.addIngestResult(result, startTime, finishTime);
-            return result;
-        });
-
-        Instant finishTaskTime = getTimeNow.get();
-        taskStatusBuilder.finishedStatus(taskFinishedStatusBuilder
-                .finish(startTaskTime, finishTaskTime).build());
-        statusStore.taskFinished(taskStatusBuilder.build());
-        LOGGER.info("IngestTask finished at = {}", finishTaskTime);
-    }
 
     public static void main(String[] args) throws IOException, StateStoreException, IteratorException, ObjectFactoryException {
         if (1 != args.length) {
@@ -123,9 +79,9 @@ public class IngestJobQueueConsumerRunner {
                 S3AsyncClient.create(),
                 HadoopConfigurationProvider.getConfigurationForECS(instanceProperties));
         IngestJobQueueConsumer queueConsumer = new IngestJobQueueConsumer(sqsClient, cloudWatchClient, instanceProperties);
-        IngestJobQueueConsumerRunner ingestJobQueueConsumerRunner = new IngestJobQueueConsumerRunner(
+        IngestTaskRunner ingestTaskRunner = new IngestTaskRunner(
                 queueConsumer, taskId, taskStore, ingestJobRunner::ingest);
-        ingestJobQueueConsumerRunner.run();
+        ingestTaskRunner.run();
 
         s3Client.shutdown();
         LOGGER.info("Shut down s3Client");
